backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
qrcode==7.4.2
Pillow>=10.3.0
python-dotenv==1.0.0
openpyxl==3.1.2
python-dateutil==2.8.2
requests==2.31.0
